@@ -1,13 +1,13 @@
 plugins {
     id 'com.android.application'
     id 'kotlin-android'
-    id 'com.google.gms.google-services'
     id 'kotlin-kapt'
     id 'com.google.dagger.hilt.android'
 }
 
 android {
-    compileSdk 34
+    compileSdk 35
+    namespace 'com.s3enterprises.jewellowholesale'
 
     defaultConfig {
         applicationId "com.s3enterprises.jewellowholesale"
@@ -25,11 +25,11 @@ android {
         }
     }
     compileOptions {
-        sourceCompatibility JavaVersion.VERSION_1_8
-        targetCompatibility JavaVersion.VERSION_1_8
+        sourceCompatibility JavaVersion.VERSION_17
+        targetCompatibility JavaVersion.VERSION_17
     }
     kotlinOptions {
-        jvmTarget = '1.8'
+        jvmTarget = '17'
     }
     buildFeatures {
         dataBinding true
@@ -39,56 +39,45 @@ android {
 
 dependencies {
 
-    implementation "org.jetbrains.kotlin:kotlin-stdlib:1.9.0"
-    implementation 'androidx.core:core-ktx:1.12.0'
-    implementation "androidx.activity:activity-ktx:1.8.0"
-    implementation "androidx.collection:collection-ktx:1.3.0"
-    implementation 'androidx.appcompat:appcompat:1.6.1'
-    implementation 'com.google.android.material:material:1.6.0'
-    implementation 'androidx.constraintlayout:constraintlayout:2.1.4'
+    implementation "org.jetbrains.kotlin:kotlin-stdlib:2.1.21"
+    implementation 'androidx.core:core-ktx:1.16.0'
+    implementation "androidx.activity:activity-ktx:1.10.1"
+    implementation "androidx.collection:collection-ktx:1.5.0"
+    implementation 'androidx.appcompat:appcompat:1.7.1'
+    implementation 'com.google.android.material:material:1.12.0'
+    implementation 'androidx.constraintlayout:constraintlayout:2.2.1'
     implementation 'androidx.legacy:legacy-support-v4:1.0.0'
-
-    testImplementation 'junit:junit:4.13.2'
-
-    //firebase
-    implementation platform('com.google.firebase:firebase-bom:27.0.0')
-    implementation 'com.google.firebase:firebase-firestore-ktx:24.0.2'
-    implementation 'com.google.firebase:firebase-auth-ktx:21.0.2'
-
 
     //ViewModel and LiveData
     implementation 'androidx.lifecycle:lifecycle-extensions:2.2.0'
-    implementation "androidx.lifecycle:lifecycle-viewmodel-ktx:2.6.2"
-    implementation "androidx.lifecycle:lifecycle-runtime-ktx:2.6.2"
-    kapt "androidx.lifecycle:lifecycle-compiler:2.6.2"
+    implementation "androidx.lifecycle:lifecycle-viewmodel-ktx:2.9.1"
+    implementation "androidx.lifecycle:lifecycle-runtime-ktx:2.9.1"
+    kapt "androidx.lifecycle:lifecycle-compiler:2.9.1"
 
     //Swipe to refresh
     implementation 'androidx.swiperefreshlayout:swiperefreshlayout:1.1.0'
 
     // Room components
-    implementation("androidx.room:room-runtime:2.6.0")
-    kapt "androidx.room:room-compiler:2.6.0"
+    implementation("androidx.room:room-runtime:2.7.1")
+    kapt "androidx.room:room-compiler:2.7.1"
     // Kotlin Extensions and Coroutines support for Room
-    implementation "androidx.room:room-ktx:2.6.0"
+    implementation "androidx.room:room-ktx:2.7.1"
 
-
-    implementation "androidx.lifecycle:lifecycle-viewmodel-ktx:2.6.2"
+    implementation "androidx.lifecycle:lifecycle-viewmodel-ktx:2.9.1"
 
     //coroutines
-    implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-android:1.7.1'
+    implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-android:1.7.3'
 
     //GSON
-    implementation 'com.google.code.gson:gson:2.9.0'
+    implementation 'com.google.code.gson:gson:2.11.0'
 
     //Rx
     implementation "io.reactivex.rxjava2:rxkotlin:2.4.0"
 
     //Hilt
-    implementation "com.google.dagger:hilt-android:2.44"
-    kapt "com.google.dagger:hilt-compiler:2.44"
+    implementation "com.google.dagger:hilt-android:2.56.2"
+    kapt "com.google.dagger:hilt-compiler:2.56.2"
 
-    androidTestImplementation 'androidx.test.ext:junit:1.1.3'
-    androidTestImplementation 'androidx.test.espresso:espresso-core:3.4.0'
 }
 
 // Allow references to generated code
